feat(poc): add an hybrid SPA/SSR interaction
In this commit, I added a `plugins/opencds/app.js` script that is a standalone,
SSR-ready Vue.js component.

This component is retrieved as a rendered raw string through a `GET
/api/v1/opencds` HTTP request made by `client` (in the user's browser)
and then loaded by Vue.js in the DOM thanks to the `v-html` directive.

Then, a dedicated `onUpdated` callback is triggred by the DOM update and this
will trigger _another render_ of the component, on the client this time, that
will be "mounted" by Vue.js on the SSR DOM node.

Thus the component is rendered by the server, loaded by the client, and Vue.js
handles all the 🪄MAGIC🪄.